+++ CHANGELOG.md
@@ -30,6 +30,7 @@ Date:
 * SPARKX checks now for static tying consistency
 * Added support for the SMASH 3.2 feature of ASCIICustom output format
 * Add option to add two Oscar/Jetscape/ParticleObjectStorer instances while preserving the event order
+* Oscar: Add function to extract the impact parameters
 
 ### Changed
 * Particle: Rename several methods for a more intuitive naming scheme. Renamed methods are: